move all scss files into new styles folder. Fix scss file references
