+++ react-front-end/src/components/Login/LoginUser.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import './LoginUser.scss';
 
 const LoginUser = () => {
   return (
