--- react-front-end/src/components/Login/LoginUser.scss
@@ -1,5 +1,0 @@
-.login-wrapper {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
