+++ react-front-end/src/components/Maps/Map.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { MapContainer, TileLayer, Marker, Popup } from 'react-leaflet'
 import L from "leaflet";
-import '../../App.scss';
+import '../../styles/App.scss';
 
 export default function MapClass (){
