+++ react-front-end/src/components/Maps/index.jsx
@@ -2,8 +2,6 @@ import React from 'react';
 import Map from './Map'
 
 import 'leaflet/dist/leaflet.css'
-import '../../App.scss';
-import '../../index.scss';
 
 export default function Maps() {
   return (
