+++ react-front-end/src/components/Meetups/index.jsx
@@ -9,8 +9,8 @@ import Chatbox from '../Chat/Chatbox'
 import ChatJoin from '../Chat/ChatJoin'
 import SafetySend from './SafetySend'
 
-import '../Button.scss';
-import '../../App.scss';
+import '../../styles/components/_button.scss'
+import '../../styles/App.scss';
 
 import { MeetupsContext } from '../../MeetupsContext.jsx';
 import { CheckedContext } from './CheckedContext.jsx'
