--- react-front-end/src/components/UserLogin/UserLogin.scss
@@ -1,5 +1,0 @@
-.login-wrapper {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
