+++ react-front-end/src/styles/App.scss
@@ -1,3 +1,14 @@
+@import './abstracts/variables';
+@import './components/button';
+@import './layout/_navbar';
+@import './components/meetups';
+@import './components/forecast';
+@import './components/loginUser';
+@import './components/_safetySend';
+@import './components/_userLogin';
+
+
+
 .App {
   text-align: center;
 }
