+++ react-front-end/src/styles/components/_forecast.scss
