+++ react-front-end/src/styles/components/_loginUser.scss
@@ -1,0 +1,5 @@
+.login-wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
