+++ react-front-end/src/styles/components/_meetups.scss
@@ -1,0 +1,3 @@
+.attendees {
+  color: red;
+}
