+++ react-front-end/src/styles/components/_safetySend.scss
