+++ react-front-end/src/styles/layout/_navbar.scss
